refactor(desktop): consolidate yt-dlp spawn/search/version helpers
Add an optional AbortSignal to utils/ytdlp-spawn.ts#spawnYtDlp and a ytSearch
helper, and move the error classifier (classifyYtDlpFailure/tailOutput/
YT_DLP_ERROR_CODES) and version utils (extractVersionSegments/hasUpdate) out
of ipc/downloader.ts into the shared util. downloader.ts re-exports them so
its test + callers are unchanged. Delete the private spawnYtDlp copy in
metadata-lookup.ts in favour of the canonical signal-aware one.
